// types/RouteType.ts
export type RouteType = {
  Home: undefined;
  HomeNext: undefined;
  Bienvenue: undefined;
  CreationEtDeveloppement: undefined;
  LoveCoach: undefined;
  LinksLogIn: {LoginRoute: string};
  LinksSignIn: {SignInRoute: string};
  Liens_de_connexion: undefined;
  S_Inscrire: undefined;
  S_inscrire_par_mail: undefined;
  S_inscrire_par_numero: undefined;
  Se_Connecter: undefined;
  Confirmation_email: undefined;
  Confirmation_numero: undefined;
  Recuperation_email: undefined;
  Ville: undefined;
  Acces_Position: undefined;
  Genre: undefined;
  Date_de_naissance: undefined;
  Taille: undefined;
  Langue_parler: undefined;
  Situation: undefined;
  Orientation: undefined;
  Recherche1: undefined;
  Recherche2: undefined;
  Affinite: undefined;
  Rythme1: undefined;
  Rythme2: undefined;
  Prenom: undefined;
  Confirmation_prenom: undefined;
  Profil_multiples: undefined;
  Prenium: undefined;
  Compte: undefined;
  Confirmation_compte: undefined;
  Recuperation_code: undefined;
  Ajouter_photo: undefined;
  Empreinte_vocal: undefined;
  Charte_engagement: undefined;
  Felicitations: undefined;
  navigateTo: string;
  Tabs: {tabPath: string};
  TabDiscover: {tabPath: string};
  TabTalk: {tabPath: string};
  TabMessages: {tabPath: string};
  TabMap: {tabPath: string};
  TabMoi: {tabPath: string};
  DiscoverNavigator: {DiscoverRoute: string};
  C_est_match: undefined;
  Discover: undefined;
  ProfilNavigator: {
    ProfilRoute: string;
    tabPath: string;
  };
  Objectifs: undefined;
  Notifications: undefined;
  SettingsStack: {
    imagePath: string | null;
    SettingsRoute: string;
  };
  Settings: undefined;
  Notifications_settings: undefined;
  Securite_et_privee: undefined;
  Contact_et_FAQ: undefined;
  Aide: undefined;
  Centre_de_securite: undefined;
  Nous_contactez: undefined;
  FAQ: undefined;
  Emplacement: undefined;
  Mode_invisible: undefined;
  Mode_voyage: undefined;
  Mettre_en_pause: undefined;
  Mode_de_connexion: undefined;
  Changer_localisation: undefined;
  Parametre_de_confidentialites: undefined;
  Autorisations_necessaires: undefined;
  Bloquer_contacts: undefined;
  Supprimer_mon_compte: undefined;
  Compte_non_trouve: undefined;
  GameNavigator: {GameRoute: string};
  Voix_du_jour: undefined;
  Carte_magique: undefined;
  AppsAffinitairesNavigator: {AppsAffinitairesRoute: string};
  Apps_Affinitaires: undefined;
  Apps_Affinitaires2: {
    routeAffinite: string;
    menu: boolean;
  };
};